Extract the logic for adding tokens to tokenList
